Nav working to the booking page, location tracking is set up and working
