+++ src/js/components/Appointment.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 
-export default class Booking extends React.Component {
+export default class Appointment extends React.Component {
 
 
   constructor(props){
