+++ src/js/components/Booking.js
@@ -1,5 +1,10 @@
 import React from "react";
-
+import BookingStore from "../stores/BookingStore";
+import BookingActions from "../actions/BookingActions";
+import Header from "../Components/MIKHeader";
+import Footer from "../Components/MIKFooter";
+import Appointment from "../Components/Appointment";
+import Checkout from "../Components/Checkout";
 
 export default class Booking extends React.Component {
 
@@ -7,7 +12,7 @@ export default class Booking extends React.Component {
   constructor(props){
     super(props);
     this.state = {
-      showCheckout:true
+      showCheckout:false
     };
   }
 
@@ -22,53 +27,22 @@ export default class Booking extends React.Component {
 
 
     return (
+      <div>
+          
+          <div>
+<div class="mod-no-checkout mod-hide">
+              <div class="m-appointments-title-bar m-tile shadow-1">
+                  <div class="m-button shadow-1 shadow-hover-2 shadow-active-3"><a href="/">View More Classes</a></div>
+                  <div class="m-button shadow-1 shadow-hover-2 shadow-active-3">Checkout</div>
+              </div>
+              
+              <div><Appointment/></div>
+              <div> <Checkout/></div>
 
+</div>
 
-        <div class="m-checkout ">
-          <div class="personal-info">
-            <div class="personal-title">Personal Information</div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">First Name</label>
-            </div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Last Name</label>
-            </div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Email</label>
-            </div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Phone Number</label>
-            </div>
           </div>
-          <div class="purchase-info">
-            <div class="purchase-title">Payment Information</div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Name</label>
-            </div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Billing Address</label>
-            </div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Credit Card</label>
-            </div>
-            <div class="m-input shadow-1">
-              <input type="text" required class="m-input-field" />
-              <label class="m-input-label">Security Code</label>
-            </div>
-          </div>
-          <div class="m-button shadow-1 shadow-hover-2 shadow-active-3 mod-book">Purchase Classes</div>
-          <div class="thank-you m-tile shadow-2 anim-tile-in">
-            <div class="thank-title">Your Class has been booked!</div>
-            <div class="thank-content">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam in turpis dapibus, tristique erat a, venenatis est. Aliquam id scelerisque metus. Nullam a maximus nibh. Curabitur auctor bibendum venenatis. Pellentesque et mi vehicula, pharetra est lobortis, mollis sem. In placerat sollicitudin eros et faucibus. Vivamus consectetur commodo egestas. Morbi pretium placerat purus, nec feugiat urna. Aliquam et interdum nibh, in feugiat nisl. In hac habitasse platea dictumst.</div>
-            <div class="m-button shadow-1 shadow-hover-2 shadow-active-3 m-button-more">View More Classes</div>
-          </div>
+          
         </div>
     );
   }
