+++ src/js/components/MIKFooter.js
@@ -1,17 +1,17 @@
 import React from "react";
 
 
-export default class Footer extends React.Component {
+export default class MIKFooter extends React.Component {
   render() {
     return (
         <div id="mik-footer" role="contentinfo">
 
-          <footer>
+          <div>
             <p>
               <a href="http://www.michaels.com/your-privacy-rights/ca-your-privacy-rights.html">Privacy Rights</a> | <a href="http://www.michaels.com/terms-and-conditions/ca-terms-and-conditions.html">Terms and Conditions</a> | <a href="http://www.michaels.com/CustomerService-ContactUs">Contact Us</a>
             </p>
             <p> 2016 Michaels Stores. Michaels and the Michaels logo and other trademarks and logos used on this site are owned or licensed by Michaels Stores, Inc. All rights reserved.</p>
-          </footer>
+          </div>
 
         </div>
     );
