+++ src/js/pages/Layout.js
@@ -1,10 +1,5 @@
 import React from "react";
 import {Link} from "react-router"
-//import './../../css/style.css';
-import Footer from "./../components/Footer";
-import Header from "./../components/Header";
-import Body from "./../components/Body";
-import { Button,ButtonToolbar } from 'react-bootstrap';
 
 export default class Layout extends React.Component {
 
@@ -24,7 +19,7 @@ export default class Layout extends React.Component {
           
 
        
-            <Link to="classesandevents/US"><button className="btn btn-lg btn-success">Go to Michaels USA</button></Link>
+            <Link to="classesandevents/US"><button >Go to Michaels USA</button></Link>
             <Link to="classesandevents/CA"><button>Go to Michaels Canada</button></Link>
